# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Movie {
  id: number;
  title: string;
  /** @nullable */
  poster_path: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  release_date: string;
  ott_platform: string;
  slug: string;
  language: string;
  /** @nullable */
  overview?: string | null;
}

export interface MovieListResponse {
  movies: Movie[];
  total_pages: number;
  page: number;
}

export interface CastMember {
  id: number;
  name: string;
  /** @nullable */
  character?: string | null;
  /** @nullable */
  profile_path?: string | null;
}

export interface MovieDetail {
  id: number;
  title: string;
  /** @nullable */
  poster_path: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count?: number;
  release_date: string;
  /** @nullable */
  runtime?: number | null;
  ott_platform: string;
  slug: string;
  language: string;
  /** @nullable */
  overview?: string | null;
  genres?: string[];
  cast?: CastMember[];
  /** @nullable */
  trailer_url?: string | null;
  /** @nullable */
  director?: string | null;
}

export interface WatchlistItem {
  id: number;
  userId: string;
  movieId: number;
  title: string;
  /** @nullable */
  poster_path: string | null;
  ott_platform: string;
  slug?: string;
  addedAt: string;
}

export interface WatchlistInput {
  userId: string;
  movieId: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  ott_platform: string;
  slug?: string;
}

export interface PaymentOrderInput {
  plan: string;
  userId: string;
}

export interface PaymentOrder {
  orderId: string;
  keyId: string;
  amount: number;
  currency: string;
}

export interface PaymentVerifyInput {
  razorpay_order_id: string;
  razorpay_payment_id: string;
  razorpay_signature: string;
  userId: string;
  plan: string;
}

export interface PaymentVerifyResult {
  success: boolean;
  message?: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface AdminStats {
  dailyUsers: number;
  monthlyRevenue: number;
  seoPages: number;
  premiumUsers: number;
  dailyUsersGrowth: string;
  revenueGrowth: string;
  seoPagesGrowth: string;
  premiumGrowth: string;
}

export interface TrafficSource {
  source: string;
  sessions: number;
  percentage: number;
}

export interface DayTraffic {
  day: string;
  visits: number;
}

export interface SeoPage {
  url: string;
  clicks: number;
  status: string;
}

export type GetTrendingMoviesParams = {
  page?: number;
  platform?: string;
};

export type SearchMoviesParams = {
  q: string;
  page?: number;
};

export type GetWatchlistParams = {
  userId: string;
};