# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  dailyUsers: number;
  monthlyRevenue: number;
  seoPages: number;
  premiumUsers: number;
  dailyUsersGrowth: string;
  revenueGrowth: string;
  seoPagesGrowth: string;
  premiumGrowth: string;
}