# lib/api-zod/src/generated/types/castMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CastMember {
  id: number;
  name: string;
  /** @nullable */
  character?: string | null;
  /** @nullable */
  profile_path?: string | null;
}