# lib/api-zod/src/generated/types/dayTraffic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DayTraffic {
  day: string;
  visits: number;
}