# lib/api-zod/src/generated/types/movieDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CastMember } from "./castMember";

export interface MovieDetail {
  id: number;
  title: string;
  /** @nullable */
  poster_path: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count?: number;
  release_date: string;
  /** @nullable */
  runtime?: number | null;
  ott_platform: string;
  slug: string;
  language: string;
  /** @nullable */
  overview?: string | null;
  genres?: string[];
  cast?: CastMember[];
  /** @nullable */
  trailer_url?: string | null;
  /** @nullable */
  director?: string | null;
}