# lib/api-zod/src/generated/types/seoPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SeoPage {
  url: string;
  clicks: number;
  status: string;
}