# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FilmDhundo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistItem {
  id: number;
  userId: string;
  movieId: number;
  title: string;
  /** @nullable */
  poster_path: string | null;
  ott_platform: string;
  slug?: string;
  addedAt: string;
}